Refactor: Extracting Menu sections into reusable components
Extract menu sections into resuable components for reuse in
HamburgerDropdownPopover.
