--- src/modules/users/components/AvatarDropdown/AvatarDropdownPopover.css
@@ -1,3 +1,0 @@
-.externalLink {
-  color: var(--grey-4);
-}
