--- src/modules/users/components/AvatarDropdown/AvatarDropdownPopover.css.d.ts
@@ -1,1 +1,0 @@
-export const externalLink: string;
