+++ src/modules/users/components/AvatarDropdown/AvatarDropdownPopover.tsx
@@ -1,59 +1,15 @@
-import React, { useCallback } from 'react';
-import { defineMessages } from 'react-intl';
+import React from 'react';
 
-import { ActionButton } from '~core/Button';
-import NavLink from '~core/NavLink';
-import ExternalLink from '~core/ExternalLink';
-import { FEEDBACK, HELP } from '~externalUrls';
-
-import { Colony } from '~data/index';
-import DropdownMenu, {
-  DropdownMenuSection,
-  DropdownMenuItem,
-} from '~core/DropdownMenu';
-import { ActionTypes } from '~redux/index';
-import {
-  USER_EDIT_ROUTE,
-  CREATE_COLONY_ROUTE,
-  CREATE_USER_ROUTE,
-} from '~routes/index';
-
-import styles from './AvatarDropdownPopover.css';
-
-const MSG = defineMessages({
-  buttonGetStarted: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.buttonGetStarted',
-    defaultMessage: 'Get started',
-  },
-  myProfile: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.myProfile',
-    defaultMessage: 'My Profile',
-  },
-  settings: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.colonySettings',
-    defaultMessage: 'Settings',
-  },
-  createColony: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.createColony',
-    defaultMessage: 'Create a Colony',
-  },
-  reportBugs: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.reportBugs',
-    defaultMessage: 'Report Bugs',
-  },
-  helpCenter: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.helpCenter',
-    defaultMessage: 'Help Center',
-  },
-  signOut: {
-    id: 'users.AvatarDropdown.AvatarDropdownPopover.link.signOut',
-    defaultMessage: 'Sign Out',
-  },
-});
+import { Colony, Maybe } from '~data/index';
+import DropdownMenu from '~core/DropdownMenu';
+import UserSection from '~users/PopoverSection/UserSection';
+import ColonySection from '~users/PopoverSection/ColonySection';
+import HelperSection from '~users/PopoverSection/HelperSection';
+import MetaSection from '~users/PopoverSection/MetaSection';
 
 interface Props {
   closePopover: () => void;
-  username?: string | null;
+  username?: Maybe<string>;
   walletConnected?: boolean;
   onlyLogout?: boolean;
   colony: Colony;
@@ -68,97 +24,18 @@ const AvatarDropdownPopover = ({
   onlyLogout = false,
   colony,
 }: Props) => {
-  const renderUserSection = useCallback(() => {
-    return (
-      <DropdownMenuSection separator>
-        {!username && (
-          <DropdownMenuItem>
-            <NavLink
-              to={{
-                pathname: CREATE_USER_ROUTE,
-                state: colony?.colonyName
-                  ? { colonyURL: `/colony/${colony?.colonyName}` }
-                  : {},
-              }}
-              text={MSG.buttonGetStarted}
-            />
-          </DropdownMenuItem>
-        )}
-        {username && (
-          <DropdownMenuItem>
-            <NavLink
-              to={`/user/${username}`}
-              text={MSG.myProfile}
-              data-test="userProfile"
-            />
-          </DropdownMenuItem>
-        )}
-        {username && (
-          <DropdownMenuItem>
-            <NavLink
-              to={USER_EDIT_ROUTE}
-              text={MSG.settings}
-              data-test="userProfileSettings"
-            />
-          </DropdownMenuItem>
-        )}
-      </DropdownMenuSection>
-    );
-  }, [colony, username]);
-
-  const renderColonySection = () => (
-    <DropdownMenuSection separator>
-      <DropdownMenuItem>
-        <NavLink to={CREATE_COLONY_ROUTE} text={MSG.createColony} />
-      </DropdownMenuItem>
-    </DropdownMenuSection>
-  );
-
-  const renderHelperSection = () => (
-    <DropdownMenuSection separator>
-      <DropdownMenuItem>
-        <ExternalLink
-          href={FEEDBACK}
-          text={MSG.reportBugs}
-          className={styles.externalLink}
-        />
-      </DropdownMenuItem>
-      <DropdownMenuItem>
-        <ExternalLink
-          href={HELP}
-          text={MSG.helpCenter}
-          className={styles.externalLink}
-        />
-      </DropdownMenuItem>
-    </DropdownMenuSection>
-  );
-
-  const renderMetaSection = () =>
-    walletConnected && (
-      <DropdownMenuSection separator>
-        <DropdownMenuItem>
-          <ActionButton
-            appearance={{ theme: 'no-style' }}
-            text={MSG.signOut}
-            submit={ActionTypes.USER_LOGOUT}
-            error={ActionTypes.USER_LOGOUT_ERROR}
-            success={ActionTypes.USER_LOGOUT_SUCCESS}
-          />
-        </DropdownMenuItem>
-      </DropdownMenuSection>
-    );
-
   return (
     <DropdownMenu onClick={closePopover}>
       {!onlyLogout ? (
         <>
-          {renderUserSection()}
-          {renderColonySection()}
-          {renderHelperSection()}
-          {renderMetaSection()}
+          {/* Move into separate components for reuse in HamburgerDropdownPopover */}
+          <UserSection colony={colony} username={username} />
+          <ColonySection />
+          <HelperSection />
+          {walletConnected && <MetaSection />}
         </>
       ) : (
-        <>{renderMetaSection()}</>
+        walletConnected && <MetaSection />
       )}
     </DropdownMenu>
   );
