+++ src/modules/users/components/PopoverSection/HelperSection.css
@@ -1,0 +1,12 @@
+@value queries: "~styles/queries.css";
+@value query700 from queries;
+
+.externalLink {
+  color: var(--grey-4);
+}
+
+@media screen and query700 {
+  .externalLink {
+    color: white;
+  }
+}
