+++ src/modules/users/components/PopoverSection/HelperSection.css.d.ts
@@ -1,0 +1,3 @@
+export const queries: string;
+export const query700: string;
+export const externalLink: string;
